# examples/smart-canteen/contracts/smart-canteen.openapi.yaml
openapi: 3.0.3
info:
  title: Smart Canteen Workflow API
  version: 1.0.0
  description: 菜单审批、采购缺口、库存入库和台账预警的最小业务闭环。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
  /api/v1/auth/refresh-token:
    post:
      operationId: refreshToken
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Rotated login tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Refresh session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /api/v1/auth/me:
    get:
      operationId: currentUser
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BusinessError:
      description: 业务规则或参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1, format: password }
        loginType: { type: string, default: account }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string, minLength: 1 }
    LogoutRequest:
      type: object
      properties:
        refreshToken: { type: string, nullable: true }
    AuthUserInfo:
      type: object
      required: [userId, username, nickname, role]
      properties:
        userId: { type: string }
        username: { type: string }
        nickname: { type: string }
        role: { type: string }
        schoolId: { type: string, nullable: true }
        canteenId: { type: string, nullable: true }
    AuthTokens:
      type: object
      required: [token, refreshToken, expiresIn, userInfo]
      properties:
        token: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, format: int64 }
        userInfo:
          $ref: '#/components/schemas/AuthUserInfo'
    CurrentUser:
      type: object
      required: [userId, username, nickname, role]
      properties:
        userId: { type: string }
        username: { type: string }
        nickname: { type: string }
        role: { type: string }
        schoolId: { type: string, nullable: true }
        canteenId: { type: string, nullable: true }
    EmptyResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
        data: { nullable: true }
    AuthTokensResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          $ref: '#/components/schemas/AuthTokens'
    CurrentUserResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          $ref: '#/components/schemas/CurrentUser'
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          nullable: true